% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducekappa-package.R
\docType{package}
\name{reducekappa-package}
\alias{reducekappa}
\alias{reducekappa-package}
\title{reducekappa: Reduce Redundant Pathway Enrichment Results Using Kappa Scores}
\description{
Clusters similar gene sets based on kappa similarity scores of shared genes, as described in the Metascape paper (Zhou et al. 2019 \doi{10.1038/s41467-019-09234-6}). Selects a representative term for each cluster and retains relevant pathway information. Supports single-trial reduction, cross-trial clustering, and comparative analysis across multiple enrichment results.
}
\author{
\strong{Maintainer}: Anne Marie Weitzel \email{weitzela@umich.edu}

}
\keyword{internal}
